requirements.txt
praw>=7.7
openai>=1.35
openai-whisper>=20231117
ffmpeg-python>=0.2.0
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
httpx>=0.27
schedule>=1.2
python-dotenv>=1.0
